requirements.txt
pandas
scikit-learn
joblib
mlflow
hydra-core
dvc
fastapi
uvicorn
streamlit
folium
streamlit-folium
shap
xgboost